Commit my local Jeri multiplexer stability improvements to assist with jtreg multiplexer nio tests:
net.jini.jeri.tcp.outOfThreads.OutOfThreads.java
net.jini.jeri.tcp.outOfThreads.OutOfThreads2.java

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/skunk/qa-refactor-namespace/trunk@1716613 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/river/jeri/internal/mux/MuxClient.java
@@ -69,7 +69,7 @@ public class MuxClient extends Mux {
 		ioe.initCause(muxDownCause);
 		throw ioe;
 	    }
-	    int sessionID = busySessions.nextClearBit(0);
+	    byte sessionID = (byte) busySessions.nextClearBit(0);
 	    if (sessionID > Mux.MAX_SESSION_ID) {
 		throw new IOException("no free sessions");
 	    }
